Added password label and confirm password field in Registration
Added a label for password to tell users it has to be more than 6
characters in length as well as adding a confirm password field with
the correct validation.

addded a unique email validation also and a notempty validation on the
password field.

also made the password fields empty themselves when the users clicks
the register button for security purposes.