Allow users to change email and password
Allowing users to change their account email and password.
Implementation is a little hacky because of the "password" validation
rules but it works, it just doesn't show an error message around the
"newPassword" or "confirmNewPassword" fields. Will fix at a later date
because I'm too tired and annoyed right now.